lang-detect/requirements.txt
fastapi
uvicorn
lingua-language-detector
